all files

+++ src/main/resources/static/js/mypage/mypage_fn.js
@@ -21,18 +21,20 @@ function myinfo_modify_p() {
 function myinfo_delete_p() {
 	var puserid = $("#puserid").val();
 	
-	$.ajax({
-		type: 'delete',
-		data: { puserid: puserid },
-		url: "/RestRegisterDelete_p",
-		success: function(responce) {
-			if(responce == 'success') {
-				window.location.href= "/main";
+	if(confirm("모든 정보가 다 삭제 됩니다. 탈퇴하시겠습니까?")) {
+		$.ajax({
+			type: 'delete',
+			data: { puserid: puserid },
+			url: "/RestRegisterDelete_p",
+			success: function(responce) {
+				if(responce == 'success') {
+					window.location.href= "/main";
+				}
+			} , error: function(xhr, status, error) {
+				console.error("Ajax 요청 실패:", status, error);
 			}
-		} , error: function(xhr, status, error) {
-			console.error("Ajax 요청 실패:", status, error);
-		}
-	});
+		});
+	}
 }
 
 function myinfo_modify_c() {
@@ -62,16 +64,18 @@ function myinfo_modify_c() {
 function myinfo_delete_c() {
 	var cuserid = $("#cuserid").val();
 	
-	$.ajax({
-		type: 'delete',
-		data: { cuserid: cuserid },
-		url: "/RestRegisterDelete_c",
-		success: function(responce) {
-			if(responce == 'success') {
-				window.location.href= "/main";
+	if(confirm("모든 정보가 삭제됩니다. 탈퇴를 진행하시겠습니까?")) {
+		$.ajax({
+			type: 'delete',
+			data: { cuserid: cuserid },
+			url: "/RestRegisterDelete_c",
+			success: function(responce) {
+				if(responce == 'success') {
+					window.location.href= "/main";
+				}
+			} , error: function(xhr, status, error) {
+				console.error("Ajax 요청 실패:", status, error);
 			}
-		} , error: function(xhr, status, error) {
-			console.error("Ajax 요청 실패:", status, error);
-		}
-	});
+		});
+	}
 }
